Initial CHERI Alliance release
This commit is to support the initial CHERI Alliance release of the CHERI model. It contains a lot of changes that are unrelated to CHERI but the CHERI code was written with them in place so we're keeping them for simplicity. The eventual plan is to upstream most of these changes to the official sail-riscv repo.